[refactor]: Refactor and update formlibrary models.
Make more use of Abstract models, remove unecessary fields, re-make migrations, update the Admin interface for the Formlibrary app.
