+++ formlibrary/models/training.py
@@ -2,9 +2,8 @@
 # -*- coding: utf-8 -*-
 
 from django.db import models
-from formlibrary.models import Service
+from .service import Service
 from workflow.models import Contact
-# from workflow.models import Program, Office, Stakeholder, Contact, Site
 
 
 class Training(Service):
@@ -12,14 +11,14 @@ class Training(Service):
     Suject to future changes: https://github.com/hikaya-io/activity/issues/421
     ? Should we edit/update the already existing TrainingAttendance, or implement from scratch?
     """
-    trainers = models.ForeignKey(
+    trainer = models.ForeignKey(
         Contact, null=True, blank=True, on_delete=models.SET_NULL)
-    training_duration = models.IntegerField(help_text="Number of days? Sessions?")
+    duration = models.IntegerField(help_text="Number of days? Sessions?")
 
     @property
     def total_individual_supported(self):
         # TODO Check all individuals, and households and their individuals
-        return 15
+        return 0
 
     # @property
     # def attendance(self):
